update first lvl: 1.add package ERROR with classes; 2.PostController was changed

+++ src/main/java/com/poryvai/blog/service/DefaultPostService.java
@@ -1,12 +1,14 @@
 package com.poryvai.blog.service;
 
 import com.poryvai.blog.entity.Post;
+import com.poryvai.blog.error.PostNotFoundException;
 import com.poryvai.blog.repository.PostRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 @Service
 public class DefaultPostService implements PostService{
@@ -25,8 +27,13 @@ public class DefaultPostService implements PostService{
     }
 
     @Override
-    public Post updatePost(Long id, Post post) {
+    public Post updatePost(Long id, Post post)throws PostNotFoundException {
         Post postDB = postRepository.findById(id).get();
+        Optional<Post> p = postRepository.findById(id);
+
+        if (!p.isPresent()) {
+            throw new PostNotFoundException("Post Id:" + id + ", Not Available");
+        }
 
         if(Objects.nonNull(post.getTitle()) && !"".equalsIgnoreCase(post.getTitle())){
             postDB.setTitle(post.getTitle());
@@ -40,7 +47,12 @@ public class DefaultPostService implements PostService{
     }
 
     @Override
-    public void deletePostById(Long id) {
+    public void deletePostById(Long id)throws PostNotFoundException {
+        Optional<Post> post = postRepository.findById(id);
+
+        if (!post.isPresent()) {
+            throw new PostNotFoundException("Post Id:" + id + ", Not Available");
+        }
         postRepository.deleteById(id);
     }
 }
